refactor: step2 문자열 계산기 패키지 명칭 변경
