+++ src/main/java/step2/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package util;
+package step2;
 
 import java.util.Arrays;
 import java.util.regex.Matcher;
