+++ src/test/java/step2/StringAddCalculatorTest.java
@@ -6,7 +6,6 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
-import util.StringAddCalculator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
